refactor: extract DiagnosticsRedactor.swift from BugNarratorDiagnostics.swift
Extracts the DiagnosticsRedactor enum (107 lines) — file-content redaction
utility used by the diagnostics export flow — into its own file.

Byte-identical move; SHA-256 b34fbc85e92fbdfc7e04db6b53b781204b6fc2387206c11f5f96d047a99ccb73
verified against origin/main:BugNarratorDiagnostics.swift lines 343-449.

Reduces BugNarratorDiagnostics.swift by 107 lines.

Tests: 667 (unchanged).
